+++ tests/aio_tests/test_async_home.py
@@ -229,6 +229,7 @@ async def test_home_getOAuthOTK(no_ssl_fake_async_home: AsyncHome):
         2018, 12, 23, 11, 38, 21, 680000
     ) + timedelta(0, utc_offset)
 
+
 @pytest.mark.asyncio
 async def test_clearconfig(no_ssl_fake_async_home: AsyncHome):
     d1 = no_ssl_fake_async_home.search_device_by_id("3014F7110000000000000001")